Ticket 666: Decrease execution time of composite rules in Beast in loaded situations
In order to decrease the execution time of file entry queries towards the DB, which is the main part of the execution time for e.g., composite rules in Beast, a file entry cache has been implemented. The cache contains UUIDs as keys and RestfulFileEntry-objects as values. If a file entry is present in the cache, i.e., has recently been added or accessed, no query towards the DB is done. The size of the cache is by default set to 0, meaning that it will not have any effect unless specifically enabled. To enable it, the bdb.xml file shall be updated with an additional constructor argument, specifying the cache size.