Ticket 692: Syslog output from bdb-server
Possibility to send log from bdb-server to syslog added. This is done by setting 'baltrad.bdb.server.log.type = syslog' in the bltnode.properties file and (optionally) specifying a log id with e.g., 'baltrad.bdb.server.log.id = mybdb-server-id'.

Also changing log level of log in backend-py which indicates that contents of arrays are ignored in function _get_attribute_sql_values. Previously this was an error-log, even though it can not be regarded as an error in most situations. Furthermore it produced large amounts of output logs in the case of using h5-files with arrays. This is now changed to a debug-level-log.