Ticket 749: Handling non-unique identifiers in /what/source
A new method for matching the correct source for a source-string has been implemented in the get_source_id-function of backend.py. With the new method, all matching key-value pairs of the database are collected and evaluated. If there are multiple source ids that have any matching key-value pair, the one that has the highest number of matches is selected.